dns/requirements.txt
requests
Pyyaml
justext
